:white_check_mark: Update test with the patterns
