+++ src/components/Header/ProjectInformations/tests/ProjectInformations.spec.tsx
@@ -1,0 +1,46 @@
+import { fireEvent, screen } from '@testing-library/react';
+
+import { render } from '@tests/helpers/render';
+
+import ProjectInformations from '@components/Header/ProjectInformations';
+import Header from '@components/Header';
+
+describe('Project Informations', () => {
+  const setOpenMock = jest.fn();
+
+  test('should rendet title name', () => {
+    render(<ProjectInformations setOpen={setOpenMock} />);
+
+    screen.getByText('Atlas de Oportunidades');
+  });
+
+  test('should render close button', async () => {
+    render(
+      <>
+        <Header isComparisonModeOn comparisonType="grid" setComparisonType={() => {}} />
+        <ProjectInformations setOpen={setOpenMock} />
+      </>,
+    );
+
+    screen.getByLabelText('close-button');
+    const menuButton = screen.getByLabelText('menu-button');
+
+    fireEvent(
+      menuButton,
+      new MouseEvent('click', {
+        bubbles: true,
+        cancelable: true,
+      }),
+    );
+  });
+
+  test('should render sections names', () => {
+    render(<ProjectInformations setOpen={setOpenMock} />);
+
+    const sections = ['Atlas de Oportunidades', 'Sobre o Atlas', 'Contribuidores'];
+
+    sections.forEach((section) => {
+      screen.getByText(section);
+    });
+  });
+});
